requirements.txt
flask>=2.3.0
flask-sqlalchemy>=3.0.0
gunicorn>=20.0.0
beautifulsoup4>=4.11.0
dnspython>=2.2.0
python-whois>=0.7.0
requests>=2.27.0
python-nmap>=0.7.1
psycopg2-binary>=2.9.0
email-validator>=1.1.0
werkzeug>=2.2.0